+++ testdata/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest/com/liferay/portal/kernel/templateparser/TemplateNode.java
@@ -9,4 +9,6 @@ public class TemplateNode extends LinkedHashMap<String, Object> {
         return null;
     }
 
+    public String getData();
+
 }