+++ changelog.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Bugfix: Resolve Portlet Names from constant in a class from the same package as the portlet class 
 - Bugfix: Fixed compatibility with IntelliJ 2019.1.x
 - Bugfix: Fixed code completion for implicit Groovy variables
+- Bugfix: TemplateNode properties (e.g. .getData()) were not offered for nested structure variables in Freemarker templates
 
 ## 0.0.4
 - Bugfix: Fixed parsing liferay-look-and-feel.xml when creating that file for the first time